requirements.txt
python-dotenv==1.0.0
google-api-python-client==2.108.0
requests==2.31.0
google-generativeai==0.3.1
openai==1.12.0
anthropic==0.18.1
beautifulsoup4==4.12.3
lxml==5.1.0
